misc/HtmlTestSuite/results/test-0941.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml">
  <head>
    <title>Date and Time Inputs</title>
    <meta content="width=device-width, maximum-scale=1.0, user-scalable=no" name="viewport"/>
    <link href="mailto:f.clari@inno-group.com" rel="author" title="Fabrice Clari"/>
    <link href="mailto:Dimitri.Bocquet@mosquito-fp7.eu" rel="author" title="Dimitri Bocquet"/>
    <link href="https://html.spec.whatwg.org/multipage/#the-input-element" rel="help"/>
    <link href="https://html.spec.whatwg.org/multipage/#dom-input-type" rel="help"/>
    <link href="https://html.spec.whatwg.org/multipage/#dom-input-value" rel="help"/>
    <link href="https://html.spec.whatwg.org/multipage/#dom-input-min" rel="help"/>
    <link href="https://html.spec.whatwg.org/multipage/#dom-input-max" rel="help"/>
    <link href="https://html.spec.whatwg.org/multipage/#dom-input-step" rel="help"/>
    <link href="https://html.spec.whatwg.org/multipage/#dom-input-stepup" rel="help"/>
    <link href="https://html.spec.whatwg.org/multipage/#dom-input-stepdown" rel="help"/>
    <script src="/resources/testharness.js"/>
    <script src="/resources/testharnessreport.js"/>
  </head>
  <body>

      <h1>Date and Time Inputs</h1>
      <div style="display: none">
        <input max="2011-12" min="2011-01" step="2" type="month" value="2011-01"/>
        <input max="2011-W50" min="2011-W20" step="2" type="week" value="2011-W40"/>
    </div>

    <div id="log">
    </div>

  <script type="text/javascript">
    test(function() {assert_equals(document.getElementsByTagName("input")[0].type, "month")}, "month type support on input element");
    test(function() {assert_equals(document.getElementsByTagName("input")[0].value, "2011-01")}, "[month] The value must be a value that is a valid global date and time string");
    test(function() {assert_equals(document.getElementsByTagName("input")[0].min, "2011-01")}, "[month] The min attribute must have a value that is a valid global date and time string");
    test(function() {assert_equals(document.getElementsByTagName("input")[0].max, "2011-12")}, "[month] The max attribute must have a value that is a valid global date and time string");
    test(function() {assert_equals(document.getElementsByTagName("input")[0].step, "2")}, "[month] The step attribute must be expressed in seconds");
    test(function() {assert_true(typeof(document.getElementsByTagName("input")[0].stepUp) == "function")}, "[month] stepUp method support on input 'month' element");
    test(function() {assert_true(typeof(document.getElementsByTagName("input")[0].stepDown) == "function")}, "[month] stepDown method support on input 'month' element");

    test(function() {assert_equals(document.getElementsByTagName("input")[1].type, "week")}, "week type support on input element");
    test(function() {assert_equals(document.getElementsByTagName("input")[1].value, "2011-W40")}, "[week] The value must be a value that is a valid global date and time string");
    test(function() {assert_equals(document.getElementsByTagName("input")[1].min, "2011-W20")}, "[week] The min attribute must have a value that is a valid global date and time string");
    test(function() {assert_equals(document.getElementsByTagName("input")[1].max, "2011-W50")}, "[week] The max attribute must have a value that is a valid global date and time string");
    test(function() {assert_equals(document.getElementsByTagName("input")[1].step, "2")}, "[week] The step attribute must be expressed in seconds");
    test(function() {assert_true(typeof(document.getElementsByTagName("input")[1].stepUp) == "function")}, "[week] stepUp method support on input 'week' element");
    test(function() {assert_true(typeof(document.getElementsByTagName("input")[1].stepDown) == "function")}, "[week] stepDown method support on input 'week' element");

  </script>

  </body>

</html>
